requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
etils==1.5.2
fastapi==0.104.1
filelock==3.13.1
flax==0.7.5
fsspec==2023.10.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.17.3
idna==3.6
importlib-resources==6.1.1
jax==0.4.20
jax-metal==0.0.4
jaxlib==0.4.14
Jinja2==3.1.2
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.3.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper==20231117
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.3
packaging==23.2
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyparsing==3.1.1
pytube==15.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rsa==4.9
safetensors==0.4.1
scipy==1.11.4
scrapetube==2.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorstore==0.1.50
tiktoken==0.5.1
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
whisper-jax @ git+https://github.com/sanchit-gandhi/whisper-jax.git@9c50a6ee5f30f6429ad46dc748603296dfb3484b
zipp==3.17.0
